==> 5. Model.ipynb <==
### Создание модели

Здесь используется мой небольшой модуль, в котором лежат методы валидации, логирование моделей, и оболочка над LGBMRegressor.

Также при запуске обучения конкретных модулей отображаемая ячейка изменится следующим образом --- будет показано на 10 итераций меньше. Это связано с тем, что раньше там стоял параметр gap=10 (для первой модели 5), который заставлял модели обучаться до тех пор, пока 10 итераций подряд не будет улучшений, и потом откатывать все к состоянии на 10 итераций назад.

Дабы не обучать эти лишние итерации, я сразу поставил нужное значение n_seeds

На всякий случай в папке submits уже лежит финальный сабмит под названием final_submit.csv, на случай если организаторы не захотят сидеть 3 часа обучая модели.

#### Примерное время обучения всех моделей - 3 часа.

import numpy as np

from sklearn.metrics import mean_squared_error  

from utils.predict import PredictorPipeline
from utils.logging import *

data = pd.read_csv("./data/prep_data.csv")

features = np.load('./data/prep_features.npy').tolist()
features_dict = np.load('./data/prep_features_dict.npy').item()

val_groups = data.loc[data.isTrain, 'group'].values
test_groups = data.loc[~data.isTrain, 'group'].values
y = data.loc[data.isTrain, 'target'].values
logger = Logger(data)

params = {
    'model': {
        'n_estimators': 100000,
        'learning_rate': 0.1,
        'max_depth': 31,
        'num_leaves': 31,
        'colsample_bytree': 0.55,
        'min_child_samples': 11,
        'n_jobs': -1
    },
    
    'validation': {
        'n_intervals': 10,
        'n_splits': 10,
        'mode': 'hierarchy'
    },
    
    'predict': {
        'predictors': [
            {'name': 'group_kkt'},
            {'name': 'group_linear'}
        ],
        'verbose': False,
        'mode': 'RegCV'
    }
}

def get_subsample(features_dict, r):
    features = []
    chosen_sections = r.choice(list(features_dict), size = 9, replace=False).tolist()
    for section in chosen_sections:
        features += features_dict[section]
    return features

logger.get_valBagger(data, features_dict, params, n_seeds=14, n_inner=6, 
                     subsample_func=get_subsample, random_state=13, verbose=True)

logger.get_valBagger(data, features_dict, params, n_seeds=19, n_inner=6, 
                     subsample_func=get_subsample, random_state=1, verbose=True)

logger.get_valBagger(data, features_dict, params, n_seeds=29, n_inner=6, 
             a        subsample_func=get_subsample, random_state=13, verbose=True)

logger.get_valBagger(data, features_dict, params, n_seeds=20, n_inner=6, 
                     subsample_func=get_subsample, random_state=8, verbose=True)

predictors = [
    {'name': 'group_kkt'},
]

logger.set_predictor(predictors)

data = pd.read_csv("./data/prep_data_old.csv")
features = np.load('./data/prep_features_old.npy').tolist()
features_dict = np.load('./data/prep_features_dict_old.npy').item()

logger.get_valBagger(data, features_dict, params, n_seeds=7, n_inner=6, 
                     subsample_func=get_subsample, random_state=19, verbose=True)

logger.get_valBagger(data, features_dict, params, n_seeds=18, n_inner=6, 
                     subsample_func=get_subsample, random_state=1, verbose=True)

predictors = [
    {'name': 'group_kkt'},
    {'name': 'group_linear'}
]

logger.set_predictor(predictors, mode='RegCV')

#logger.save('./logger_10_1.pickle')

submit = pd.DataFrame(logger.get_test_preds(), columns=['target'])

submit.to_csv('./data/submits/submit.csv')
